Refactor and add classes to enable parsing ChordPart/MelodyPart to data sequence
To allow this project to use Impro-Visor classes, this adds a built improvisor.jar
file to the build process, and removes some conflicting classes. It also renames
Chord to ChordInfo to prevent conflicts.

The new DataPartIO class serves a similar function to LeadSheetIO, except that
it is designed to work with ChordPart(s) and MelodyPart(s).

Since Chord objects in improvisor have embedded note information, we can simply
grab that from the chord directly, instead of having to parse the chord string
ourselves. The old method is still included for backwards compatibility.