Replaces old loading system with a tree based approach. Network pieces implementing loadable must construct a node tree that has access to all its weights and matrices. Although the load method can be overwritten for any loadable, there is now a default load method which searches the node tree until it finds either another loadable (which it will call load on - that network piece can control what happens next) or until it finds a leaf node or the target of the load path, marking nodes as succesful when loaded.
The connectome loader now uses this tree structure to generate a list of
unloaded path names that exist in the network's architecture. Driver
utilizes this to supply more informative error messages. Further
development could possibly use this tree structure for other analysis, as
it provides access to all connectome data in any structure of loadables.